requirements.txt
asgiref==3.11.0
chardet==5.2.0
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
Django==5.2.10
django-cors-headers==4.9.0
django-environ==0.11.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
et_xmlfile==2.0.0
fonttools==4.61.1
fuzzywuzzy==0.18.0
kiwisolver==1.4.9
Levenshtein==0.27.3
matplotlib==3.10.8
num2words==0.5.14
numpy==2.4.1
openpyxl==3.1.5
packaging==26.0
pandas==3.0.0
pillow==12.1.0
PyJWT==2.10.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-Levenshtein==0.27.3
RapidFuzz==3.14.3
reportlab==4.4.9
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
xlrd==2.0.2

# Celery & Redis
celery[redis]>=5.3.0
redis>=5.0.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0
psycopg2-binary>=2.9.0